Update the way files are added in the gemspec
Since the gem is packaged based on this file list, it seems pertinent to
only include files that users might want to see when they install the
gem, so let's explicitly list those files from the root that we want to
include.

Also, let's use Ruby globbing instead of `git ls-files` so it's all
in-Ruby calls.

Note: `.yardopts` is included explicitly to show the file list on
rdoc.info if later added. See [this Twitter conversation][twitter].

[twitter]: https://twitter.com/lsegal/status/270950893875245057