Cache bundler dependencies on Travis
Along with the Docker enhancement, this should speed up our build time
by quite a bit. It's been verified on several open source Ruby projects,
so shouldn't be an issue.

The main bottleneck for our CI now is that Rubinius takes a while to
install from RVM. I'm not sure if there's anything we can do about that,
but it is an order of magnitude slower than the rest of the tests.