Add a conditional check for 2.2.x specs
This adds a way to handle the broken spec in MRI 2.2.x that was
introduced by a regression in the language. It is slated to be in 2.3.0
and seems partially fixed in 2.2.2.

In order to have the full spec suite run on every version of Ruby but
keep the brevity of the Mash specs, we needed a way to check for the
Ruby version and selectively disable the two errant specs.

We need to be testing on the latest Ruby, so this seems to be the best
compromise.

For more information on the breakage in Ruby, see [issue 285][rubybug].

For more information on this decision, see [issue 294][workaround].

Fixes #294
/cc #285

[rubybug]: https://github.com/intridea/hashie/pull/285
[workaround]: https://github.com/intridea/hashie/pull/294