Ensure that MethodQuery methods return booleans.
The documentation about the MethodQuery module spells out the expected
behavior. While the spec suite appeared to be testing the same, the
RSpec matchers were not actually checking the behavior properly. This
lead to a divergence between the expected behavior, as outlined in the
module documentation, and the actual behavior, as tested in the spec
suite.

This is the reason that #299 is happening: the expected behavior is not
the actual behavior. I have included a test for #299 to show that it
works as expected in this branch.

This change makes the spec suite match the behavior in the module
documentation, then modifies the methods to match the documentated
specification.

Lastly, I took the time to clean up the MethodQuery module for two
reasons:

1. The old implementation used Regexps on the stringified
   method names, which is unnecessary since we have the wonderful
   `String#end_with?` method to check the suffix of a strings.
2. The old logic was difficult to reason through, so I clarified the
   intent of the methods by reducing the complexity of the conditionals
   and extracting meaningful methods where possible.