Add documentation for Mash subclasses
The behavior of Mash subclasses around the wrapping of inner sub-Hashes can be
confusing, so we should have some documentation around it. This captures the
process of subclasses wrapping their sub-Hashes in a guided fashion.

Also, this reorganizes a bit of the Mash readme to group related topics together
under headlines.