Allow codependent Dash attributes to initialize
Definition: Codependent properties
A set of two or more properties who have "required" validations that are based
on each other.

Example:

```ruby
class OneOrMore < Hashie::Dash
  property :a, required: -> { b.nil? }
  property :b, required: -> { a.nil? }
end
```

When constructing a Dash via the merge initializer, codependent properties have
their "required" validation run too early when their values are set to `nil`,
which causes an `ArgumentError` to be raised in the case that the first property
is set to `nil`.

This is an order-dependence bug that is fixed by this commit. By pruning off
`nil` values only during initialization via the merge initializer, we can
prevent this problem from occurring for the case of `nil` values.

However, this is an indication of a larger problem with the architecture of
Dash. We should be setting all the properties before running the validations.
Rearchitecting this will be quite an undertaking.