Drop testing for unsupported Rubies
Per the [JRuby security policy][1], JRuby 9.2 is supported, but 9.0 is
not any longer.

JRuby 9.0 has been flaky on builds, so it's now causing a maintenance
burden.

This isn't a declaration that we won't support this version, but I
don't think it makes sense to test on them any more. In the next major
release, I would like to define a support policy.

[1]: https://github.com/jruby/jruby/blob/4fd5e619f9d1f36d7bbca8f0013bb55e4ad57e8f/SECURITY.md